⚡ :: 삭제 구별 구현
