+++ src/recoil/userAtom.js
@@ -1,0 +1,6 @@
+import { atom } from "recoil";
+
+export const userStateAtom = atom({
+  key: "userState",
+  default: {},
+});
